update game play system

+++ Assets/Scripts/ViewManager/ViewConfig.cs
@@ -37,7 +37,7 @@ public class ShopViewParam : ViewParam
 public class MainScreenViewParam : ViewParam
 {
     public long totalGold;
-    internal int ticket;
+    internal long ticket;
     internal int level;
 }
 public class CollectionParam : ViewParam
